requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.1=cpu_0
absl-py=0.13.0=py38h06a4308_0
aiohttp=3.7.4=py38h27cfd23_1
ase=3.22.0=pyhd8ed1ab_0
async-timeout=3.0.1=py38h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
boto=2.49.0=py38_0
boto3=1.17.109=pyhd3eb1b0_0
botocore=1.20.109=pyhd3eb1b0_1
brotlipy=0.7.0=py38h27cfd23_1003
bz2file=0.98=py38h06a4308_1
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.7.5=h06a4308_1
cached-property=1.5.2=py_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chardet=3.0.4=py38h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
coverage=5.5=py38h27cfd23_2
cryptography=3.4.7=py38hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py38_0
cython=0.29.24=py38h295c915_0
dbus=1.13.18=hb2f20db_0
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
flask=1.1.2=pyhd3eb1b0_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pyhd3eb1b0_0
future=0.18.2=py38_1
gensim=3.8.3=py38h2531618_2
glib=2.69.0=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
googledrivedownloader=0.4=pyhd3deb0d_1
grpcio=1.36.1=py38h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.2.1=py38h6c542dc_0
hdf5=1.10.6=hb1b8bf9_0
html5lib=1.1=py_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py38h06a4308_0
intel-openmp=2019.4=243
isodate=0.6.0=py_1
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.0.1=pyhd3eb1b0_0
jmespath=0.10.0=py_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
keepalive=0.5=py38h578d9bd_5
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libmklml=2019.0.5=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvmlite=0.36.0=py38h612dafd_4
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py38h06a4308_0
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.3.4=py38h06a4308_0
matplotlib-base=3.3.4=py38h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
multidict=5.1.0=py38h27cfd23_2
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
networkx=2.6.1=pyhd3eb1b0_0
ninja=1.10.2=hff7bd54_1
nltk=3.6.2=pyhd3eb1b0_0
numba=0.53.1=py38ha9443f7_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h7f98852_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.2.4=py38h2531618_0
pcre=8.45=h295c915_0
pillow=8.3.1=py38h2c7a002_0
pip=21.1.3=py38h06a4308_0
plyfile=0.7.4=pyhd8ed1ab_0
protobuf=3.17.2=py38h295c915_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pydeprecate=0.3.1=pyhd8ed1ab_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38h06a4308_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pyhd3eb1b0_0
python-levenshtein=0.12.2=py38h497a2fe_0
python-louvain=0.15=pyhd3deb0d_0
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-cluster=1.5.9=py38_torch_1.8.0_cu102
pytorch-geometric=1.7.2=py38_torch_1.8.0_cu102
pytorch-lightning=1.4.0=pyhd8ed1ab_0
pytorch-scatter=2.0.7=py38_torch_1.8.0_cu102
pytorch-sparse=0.6.10=py38_torch_1.8.0_cu102
pytorch-spline-conv=1.2.1=py38_torch_1.8.0_cu102
pytorch_cluster=1.5.4=py38h9235441_1
pytorch_geometric=1.6.1=pyh9f0ad1d_0
pytorch_scatter=2.0.4=py38h9235441_1
pytorch_sparse=0.6.3=py38h9235441_1
pytorch_spline_conv=1.2.0=py38h9235441_1
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py38h27cfd23_1
qt=5.9.7=h5867ecd_1
rdflib=6.0.0=py38h578d9bd_0
readline=8.1=h27cfd23_0
regex=2021.7.6=py38h7f8727e_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.4.2=pyhd3eb1b0_0
scikit-learn=0.24.2=py38ha9443f7_0
scipy=1.6.2=py38h91f5cce_0
setuptools=52.0.0=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
smart_open=1.9.0=py_0
sparqlwrapper=1.8.5=py38h578d9bd_1005
sqlite=3.36.0=hc218d9a_0
tbb=2020.3=hfd86e86_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
threadpoolctl=2.2.0=pyhb85f177_0
tk=8.6.10=hbc83047_0
torchaudio=0.8.0=py38
torchmetrics=0.4.1=pyhd8ed1ab_0
torchvision=0.9.0=py38_cu102
tornado=6.1=py38h27cfd23_0
tqdm=4.61.2=pyhd3eb1b0_1
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
